+++ src/main/java/nextstep/subway/line/LineRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.line;
 
 public class LineRequest {
     private String name;
