+++ src/main/java/nextstep/subway/station/StationRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.station;
 
 public class StationRequest {
     private String name;
