+++ src/main/java/nextstep/subway/station/StationResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.station;
 
 public class StationResponse {
     private Long id;
