+++ src/main/java/nextstep/subway/station/StationService.java
@@ -1,9 +1,5 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.station;
 
-import nextstep.subway.applicaion.dto.StationRequest;
-import nextstep.subway.applicaion.dto.StationResponse;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
