+++ src/main/java/nextstep/subway/ui/LineController.java
@@ -1,9 +1,9 @@
 package nextstep.subway.ui;
 
-import nextstep.subway.applicaion.LineService;
-import nextstep.subway.applicaion.dto.LineRequest;
-import nextstep.subway.applicaion.dto.LineResponse;
-import nextstep.subway.applicaion.dto.SectionRequest;
+import nextstep.subway.line.LineService;
+import nextstep.subway.line.LineRequest;
+import nextstep.subway.line.LineResponse;
+import nextstep.subway.section.SectionRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
